// src/lib/reward-api/models/data_ApproveIssueRequestRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type data_ApproveIssueRequestRequest = {
    remark?: string;
    status: data_ApproveIssueRequestRequest.status;
};
export namespace data_ApproveIssueRequestRequest {
    export enum status {
        APPROVED = 'APPROVED',
        REJECTED = 'REJECTED',
    }
}